requirements.txt
tensorflow==2.16.1
numpy
matplotlib
scikit-learn
pillow
streamlit
seaborn